requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
tqdm>=4.66.4
langchain>=0.2.11
langchain-community>=0.2.10
langchain-openai>=0.1.17
faiss-cpu>=1.8.0
pypdf>=4.2.0
docx2txt>=0.8
sentence-transformers>=3.0.1
